requirements.txt
scikit-learn==1.3.2
pandas==2.0.3
numpy==1.24.4
torch==1.8.1
tqdm==4.66.4
PyYAML==3.13
opencv-python==4.1.2.30
opencv-python-headless==4.10.0.82
albumentations==1.4.8
glob2==0.7
matplotlib==3.7.5
matplotlib-inline==0.1.7
argparse==1.1
#following packges are required if you are running jupyter notebook
ipykernel==6.29.4
ipython==8.12.3